Change color list to byte

+++ Models/SkeletonGame.Models/Text.cs
@@ -6,7 +6,7 @@ namespace SkeletonGame.Models
     public class Text
     {
         [YamlMember(Alias = "color", ApplyNamingConventions = false)]
-        public List<int> Color { get; set; }
+        public List<byte> Color { get; set; }
 
         [YamlMember(Alias = "y_center", ApplyNamingConventions = false)]
         public float YCenter { get; set; }
